Add external OAuth support
