+++ src/test/java/net/snowflake/ingest/connection/MockOAuthClient.java
@@ -2,24 +2,28 @@ package net.snowflake.ingest.connection;
 
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicReference;
+import net.snowflake.client.jdbc.internal.apache.http.client.utils.URIBuilder;
 import net.snowflake.ingest.utils.ErrorCode;
 import net.snowflake.ingest.utils.SFException;
 
 /** Mock implementation of {@link OAuthClient}, only use for test */
-public class MockOAuthClient implements OAuthClient {
+public class MockOAuthClient extends OAuthClient {
   private final AtomicReference<OAuthCredential> oAuthCredential;
 
   private int futureRefreshFailCount = 0;
 
   public MockOAuthClient() {
-    OAuthCredential mockOAuthCredential =
-        new OAuthCredential("CLIENT_ID", "CLIENT_SECRET", "REFRESH_TOKEN");
-    oAuthCredential = new AtomicReference<>(mockOAuthCredential);
+    super(
+        "ACCOUNT_NAME",
+        new OAuthCredential("CLIENT_ID", "CLIENT_SECRET", "REFRESH_TOKEN"),
+        new URIBuilder());
+    oAuthCredential =
+        new AtomicReference<>(new OAuthCredential("CLIENT_ID", "CLIENT_SECRET", "REFRESH_TOKEN"));
     oAuthCredential.get().setExpiresIn(600);
   }
 
   @Override
-  public AtomicReference<OAuthCredential> getoAuthCredentialRef() {
+  public AtomicReference<OAuthCredential> getOAuthCredentialRef() {
     return oAuthCredential;
   }
